refactor(tests): resolve Codacy issues in UserServiceProfile tests
- Remove 632-line comprehensive test file exceeding 500-line limit
- Split into 5 focused test files, each under 500 lines
- Extract helper functions to eliminate 24 code clones
- Create reusable MockServiceHelpers, AssertionHelpers, and TestScenarios
- Maintain 98.43% test coverage with 40 passing tests
- All ESLint checks pass with 0 errors/warnings

Files created:
- UserServiceProfile.test-helpers.ts (253 lines) - Helper utilities
- UserServiceProfile.updates.test.ts (242 lines) - Profile update tests
- UserServiceProfile.deletion-helpers.test.ts (215 lines) - Deletion tests
- UserServiceProfile.retrieval.test.ts (153 lines) - Retrieval tests
- UserServiceProfile.subscriptions.test.ts (97 lines) - Subscription tests

Resolves code quality issues:
✅ File length: 632 lines → max 253 lines per file
✅ Code clones: 24 instances → 0 instances
✅ Test coverage: Maintained at 98.43%
✅ All tests passing: 40/40
